app - move translation-specific work into seprate Translator class
